Allow overriding the iOS platform destination for RN E2E testing
The current setup works because the resulting destination is unique in
the CircleCI image. However, in the upcoming Buildkite setup the value
does not work and `xcodebuild` tries to build using the VM as the
destination, which fails.

See
https://buildkite.com/automattic/gutenberg-mobile/builds/7039#018a4555-02d8-4ce3-b463-b876b3e7a658/858-883

With an override-via-env-var in place, we'll be able to customize the
destination used in CI without having to further update the code here.